UINavigationController Extension file is added and the codes are in the right place
