+++ TellingMe/tellingMe/util/extension/UINavigationController+.swift
@@ -5,4 +5,15 @@
 //  Created by KYUBO A. SHIM on 2023/10/23.
 //
 
-import Foundation
+import UIKit
+
+extension UINavigationController: UIGestureRecognizerDelegate {
+    open override func viewDidLoad() {
+        super.viewDidLoad()
+        interactivePopGestureRecognizer?.delegate = self
+    }
+    
+    public func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
+        return viewControllers.count > 1
+    }
+}
