+++ TellingMe/tellingMe/util/extension/UIViewController+.swift
@@ -64,14 +64,3 @@ extension UIViewController {
         })
     }
 }
-
-extension UINavigationController: UIGestureRecognizerDelegate {
-    open override func viewDidLoad() {
-        super.viewDidLoad()
-        interactivePopGestureRecognizer?.delegate = self
-    }
-    
-    public func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
-        return viewControllers.count > 1
-    }
-}
